refactor(settings/ai-providers): extract SimpleModeSection
Move SimpleModeSection and its helpers (filterImageModels,
filterWebResearchModels, TIER_ROWS, SimpleModeModelRow, AutosaveStatus)
from the monolithic org-ai-providers.tsx into a dedicated
ai-providers/simple-mode-section.tsx file. No behavior change.
